feat: gas optimizations, move tests from DecayCalculations to the PresenceToken tests file
